[NEMO-398] ExecutorRepresenter interface and DefaultExecutorRepresenter (#227)
JIRA: [NEMO-398: ExecutorRepresenter interface and LambdaExecutorRepresenter ](https://issues.apache.org/jira/projects/NEMO/issues/NEMO-398)

**Major Changes:**
`DefaultExecutorRepresenter` has the same function as previous `ExecutorRepresenter` does.
Now `DefaultExecutorRepresenter` implements `ExecutorRepresenter` interface.

**Tests for the changes:**
- Not any
