+++ runtime/master/src/main/java/org/apache/nemo/runtime/master/resource/ContainerManager.java
@@ -193,7 +193,8 @@ public final class ContainerManager {
 
     // Create the executor representation.
     final ExecutorRepresenter executorRepresenter =
-      new ExecutorRepresenter(executorId, resourceSpec, messageSender, activeContext, serializationExecutorService,
+      new DefaultExecutorRepresenter(executorId, resourceSpec, messageSender,
+        activeContext, serializationExecutorService,
         activeContext.getEvaluatorDescriptor().getNodeDescriptor().getName());
 
     requestLatchByResourceSpecId.get(resourceSpec.getResourceSpecId()).countDown();
